### 2024-09-16 #7231 @editorialbot
[PRE REVIEW]: SlicerSPECTRecon: A 3D Slicer Extension for SPECT Image Reconstruction
**Submitting author:** @ObedDzik (<a href="http://orcid.org/0000-0002-1122-0629">Obed Dzikunu</a>)
**Repository:** https://github.com/PyTomography/slicer_spect_recon.git
**Branch with paper.md** (empty if default branch): j_paper
**Version:** v1.0.0
