+++ paper/paper.md
@@ -8,7 +8,7 @@ tags:
   - image reconstruction
 authors:
   - name: Obed K. Dzikunu
-    orcid: 0000-0002-1162-0629
+    orcid: 0000-0002-1122-0629
     corresponding: true # (This is how to denote the corresponding author)
     affiliation: "2, 3" # (Multiple affiliations must be quoted)
   - name: Maziar Sabouri
