### 2024-10-16 4f584e9
add joint multi peak, uncertainty estimation, and rearrange/rename modules
