+++ SlicerSPECTRecon/Logic/priors.py
@@ -1,6 +1,6 @@
 from pytomography.io.SPECT import dicom
 from pytomography.priors import RelativeDifferencePrior, QuadraticPrior, LogCoshPrior, TopNAnatomyNeighbourWeight
-from Logic.vtkkmrmlutils import *
+from Logic.VtkkmrmlUtils import *
 
 def selectPrior(prior_type, prior_beta, prior_delta, prior_gamma, files_NM, bed_idx, 
                 index_peak, N_prior_anatomy_nearest_neighbours, prior_anatomy_image_file):
@@ -14,20 +14,11 @@ def selectPrior(prior_type, prior_beta, prior_delta, prior_gamma, files_NM, bed_
         else:
             prior_weight = None
         if prior_type=='RelativeDifferencePenalty':
-            prior = relativeDifferencePenalty(prior_beta=prior_beta, prior_gamma=prior_gamma, prior_weight=prior_weight)
+            prior = RelativeDifferencePrior(beta=prior_beta, gamma=prior_gamma, weight=prior_weight)
             return prior
         elif prior_type=='Quadratic':
-            prior = quadraticPrior(prior_beta=prior_beta, prior_delta=prior_delta, prior_weight=prior_weight)
+            prior = QuadraticPrior(beta=prior_beta, delta=prior_delta, weight=prior_weight)
             return prior
         elif prior_type=='LogCosh':
-            prior = logCoshPrior(prior_beta=prior_beta, prior_delta=prior_delta, prior_weight=prior_weight)
+            prior = LogCoshPrior(beta=prior_beta, delta=prior_delta, weight=prior_weight)
             return prior
-
-def relativeDifferencePenalty(prior_beta, prior_gamma, prior_weight):
-    return RelativeDifferencePrior(beta=prior_beta, gamma=prior_gamma, weight=prior_weight)
-
-def quadraticPrior(prior_beta, prior_delta, prior_weight):
-    return QuadraticPrior(beta=prior_beta, delta=prior_delta, weight=prior_weight)
-
-def logCoshPrior(prior_beta, prior_delta, prior_weight):
-    return LogCoshPrior(beta=prior_beta, delta=prior_delta, weight=prior_weight)
