+++ SlicerSPECTRecon/SlicerSPECTRecon.py
@@ -4,13 +4,14 @@ import importlib
 def ensure_packages_installed():
     """Ensure required packages are installed."""
     required_packages = {
-        "pytomography": "3.2.2",
+        "pytomography": "3.2.3",
         "beautifulsoup4": None,
     }
     for package, version in required_packages.items():
         try:
             if version:
                 package_name = f"{package}=={version}"
+                slicer.util.pip_install(package_name)
             else:
                 package_name = package
             importlib.import_module(package)
@@ -36,11 +37,11 @@ from slicer import vtkMRMLScalarVolumeNode
 from DICOMLib import DICOMUtils
 import re
 from Logic.SlicerSPECTReconLogic import SlicerSPECTReconLogic
-from Logic.SlicerSPECTReconTest import SlicerSPECTReconTest
-from Logic.vtkkmrmlutils import *
-from Logic.getmetadatautils import *
-from Logic.simindToDicom import *
-from Logic.simindToDicomConverterTest import simindToDicomConverterTest
+from Logic.VtkkmrmlUtils import *
+from Logic.MetadataUtils import *
+from Logic.SimindToDicom import *
+from Testing.Python.reconstruction import ReconstructionTest
+from Testing.Python.simind_to_dicom import SimindToDicomConverterTest
 
 __submoduleNames__ = [
     "SlicerSPECTReconLogic",
@@ -96,34 +97,24 @@ class SlicerSPECTRecon(ScriptedLoadableModule):
             slicer.selfTests = {}
         slicer.selfTests["SlicerSPECTReconTest"] = self.runTest
 
-    def runTest(self, msec=100, **kwargs):
-            """
-            :param msec: delay to associate with :func:`ScriptedLoadableModuleTest.delayDisplay()`.
-            """
-            logging.info("\n******* Starting Tests of SlicerSPECTRecon **********\n")
-            # Test reconstructSimind
-            testCase = simindToDicomConverterTest()
-            testCase.messageDelay = msec
-            testCase.runTest(**kwargs)
-            # Test SlicerSPECTReconTest
-            # name of the test case class is expected to be <ModuleName>Test
-            module = importlib.import_module(self.__module__)
-            className = self.moduleName + "Test"
-            try:
-                TestCaseClass = getattr(module, className)
-            except AttributeError:
-                # Treat missing test case class as a failure; provide useful error message
-                raise AssertionError(
-                    f"Test case class not found: {self.__module__}.{className} "
-                )
-            testCase = TestCaseClass()
-            testCase.messageDelay = msec
-            testCase.runTest(**kwargs)
-            logging.info("\n******* All tests passed **********\n")
+    def runTest(self, msec=1000, **kwargs):
+        """
+        :param msec: delay to associate with :func:`ScriptedLoadableModuleTest.delayDisplay()`.
+        """
+        logging.info("\n******* Starting Tests of SlicerSPECTRecon **********\n")
+        # Test SIMIND Converter
+        testCase = SimindToDicomConverterTest()
+        testCase.messageDelay = msec
+        testCase.runTest(**kwargs)
+        # # Test Reconstructions
+        testCase = ReconstructionTest()
+        testCase.messageDelay = msec
+        testCase.runTest(**kwargs)
+        logging.info("\n******* All tests passed **********\n")
 
[…]

[…]
