+++ SlicerSPECTRecon/Testing/Python/reconstruction.py
@@ -1,0 +1,622 @@
+import os
+import slicer
+from slicer.i18n import tr as _
+from slicer.i18n import translate
+from slicer.ScriptedLoadableModule import *
+from DICOMLib import DICOMUtils
+from pytomography.io.SPECT import dicom
+import numpy as np
+import pydicom
+import torch
+import ctk
+import zipfile
+from Logic.SlicerSPECTReconLogic import SlicerSPECTReconLogic
+from Logic.VolumeUtils import *
+from Logic.Algorithms import *
+from Logic.Priors import *
+from Logic.VtkkmrmlUtils import *
+from Logic.MetadataUtils import *
+from Logic.SimindToDicom import *
+import json
+from Testing.Python.utils import *
+
+class ReconstructionTest(ScriptedLoadableModuleTest):
+    """
+    This is the test case for your scripted module.
+    Uses ScriptedLoadableModuleTest base class, available at:
+    https://github.com/Slicer/Slicer/blob/main/Base/Python/slicer/ScriptedLoadableModule.py
+    """
+
+    def setUp(self):
+        """Do whatever is needed to reset the state - typically a scene clear will be enough."""
+        self.cleanUP()
+        logger.info("SlicerSPECTReconTests")
+        module_dir = os.path.dirname(__file__)
+        self.resources_dir = os.path.join(module_dir, './../Resources')
+        self.logic = SlicerSPECTReconLogic()
+        try:
+            self.delayDisplay("Initializing Dicom Database")
+            initDICOMDatabase()
+        except:
+            raise
+        try:
+            self.uploadtoDB()
+        except:
+            raise
+
+    def cleanUP(self):
+        slicer.mrmlScene.Clear(0)
+        slicer.app.processEvents()
+
+    def runTest(self):
+        """Run as few or as many tests as needed here."""
+        self.setUp()
+        self.test_load_projection_data()
+        self.test_projection_metadata()
+        self.test_attenuation_map_alignment()
+        self.test_psf_metadata()
+        self.test_scatter()
+        self.run_test_osem()
+        self.run_test_bsrem()
+        self.run_test_osmaposl()
+        self.cleanUP()
+
+    def importDICOMFromURL(self, file_id, folder_name):
+        dicomDir = Path(slicer.app.temporaryPath) /folder_name
+        dicomDir.mkdir(parents=True, exist_ok=True)
+        zip_path = dicomDir / 'dicom.zip'
+        logger.info(f"Downloading {folder_name} DICOM files from drive to {dicomDir}")
+        status = download_file_from_google_drive(file_id, zip_path)
+        if status:
+            with zipfile.ZipFile(zip_path, 'r') as zip_ref:
+                zip_ref.extractall(dicomDir)
+        else:
+            logger.error(f"Failed to download {folder_name} DICOM files")
+            return
+        indexer = ctk.ctkDICOMIndexer()
+        indexer.addDirectory(slicer.dicomDatabase, str(dicomDir))
+        indexer.waitForImportFinished()
+        logger.info(f"Imported DICOM files from drive to database")
+
+    def downloadAndImportNMFiles(self):
+        file_id = "1bCz_hLgASAiQ38QrRlgrJ3lH_lOlqQb1"
+        folder_name = "NM"
+        self.importDICOMFromURL(file_id, folder_name)
+
+    def downloadAndImportCTFiles(self):
+        file_id = "1kbl4nqflovxOB0N9z8YrpmZHzTnPtnyu"
+        folder_name = "CT"
+        self.importDICOMFromURL(file_id, folder_name)
+    
[…]
