### 2024-10-23 #7399 @mstimberg
There are additional guidelines in the [message at the start of this issue](https://github.com/openjournals/joss-reviews/issues/7399#issue-2607617891).
