### 2024-10-24 #7399 @zapaishchykova
Hello!
I downloaded the example (- url to sample SPECT file
https://drive.google.com/file/d/1bCz_hLgASAiQ38QrRlgrJ3lH_lOlqQb1/view?usp=sharing) and trying to reconstruct, see below the error log that I'm seeing:

```
Traceback (most recent call last):
  File "/Users/anna/Downloads/slicer_spect_recon-development/SlicerSPECTRecon/SlicerSPECTRecon.py", line 474, in onReconstructButton
    print(recon_volume_node.GetID())
AttributeError: 'NoneType' object has no attribute 'GetID'
[Python] Please select a photopeak energy window
Traceback (most recent call last):
  File "/Users/anna/Downloads/slicer_spect_recon-development/SlicerSPECTRecon/SlicerSPECTRecon.py", line 453, in onReconstructButton
    recon_volume_node = self.logic.reconstruct(
  File "/Users/anna/Downloads/slicer_spect_recon-development/SlicerSPECTRecon/Logic/SlicerSPECTReconLogic.py", line 128, in reconstruct
    volume_node = self.create_volume_node_from_recon(reconstructed_image_multibed, files_NM)
  File "/Users/anna/Downloads/slicer_spect_recon-development/SlicerSPECTRecon/Logic/SlicerSPECTReconLogic.py", line 170, in create_volume_node_from_recon
    saveFilesInBrowser(temp_file_path)
  File "/Users/anna/Downloads/slicer_spect_recon-development/SlicerSPECTRecon/Logic/VolumeUtils.py", line 11, in saveFilesInBrowser
    dicomBrowser = slicer.modules.DICOMWidget.browserWidget.dicomBrowser
AttributeError: module 'modules' has no attribute 'DICOMWidget'
Traceback (most recent call last):
  File "/Users/anna/Downloads/slicer_spect_recon-development/SlicerSPECTRecon/SlicerSPECTRecon.py", line 453, in onReconstructButton
    recon_volume_node = self.logic.reconstruct(
  File "/Users/anna/Downloads/slicer_spect_recon-development/SlicerSPECTRecon/Logic/SlicerSPECTReconLogic.py", line 128, in reconstruct
    volume_node = self.create_volume_node_from_recon(reconstructed_image_multibed, files_NM)
  File "/Users/anna/Downloads/slicer_spect_recon-development/SlicerSPECTRecon/Logic/SlicerSPECTReconLogic.py", line 170, in create_volume_node_from_recon
    saveFilesInBrowser(temp_file_path)
  File "/Users/anna/Downloads/slicer_spect_recon-development/SlicerSPECTRecon/Logic/VolumeUtils.py", line 11, in saveFilesInBrowser
    dicomBrowser = slicer.modules.DICOMWidget.browserWidget.dicomBrowser
AttributeError: module 'modules' has no attribute 'DICOMWidget'
```
FYI using the dev brunch
