### 2024-10-25 #7399 @lukepolson
Sounds good, we also have a [user manual](https://github.com/PyTomography/slicer_spect_recon/blob/development/User_Manual.md) file as well, but maybe it is difficult to find, so we can make it more explicit in the README.
