### 2024-10-29 #7399 @zapaishchykova
Hi @lukepolson,

No I was not able to reconstruct the image, I used the dev brach, copying the error trace again below:
```
Traceback (most recent call last):
  File "/Users/anna/Downloads/slicer_spect_recon-development/SlicerSPECTRecon/SlicerSPECTRecon.py", line 453, in onReconstructButton
    recon_volume_node = self.logic.reconstruct(
  File "/Users/anna/Downloads/slicer_spect_recon-development/SlicerSPECTRecon/Logic/SlicerSPECTReconLogic.py", line 128, in reconstruct
    volume_node = self.create_volume_node_from_recon(reconstructed_image_multibed, files_NM)
  File "/Users/anna/Downloads/slicer_spect_recon-development/SlicerSPECTRecon/Logic/SlicerSPECTReconLogic.py", line 170, in create_volume_node_from_recon
    saveFilesInBrowser(temp_file_path)
  File "/Users/anna/Downloads/slicer_spect_recon-development/SlicerSPECTRecon/Logic/VolumeUtils.py", line 11, in saveFilesInBrowser
    dicomBrowser = slicer.modules.DICOMWidget.browserWidget.dicomBrowser
AttributeError: module 'modules' has no attribute 'DICOMWidget'
```
 Let me know if thats something that I'm clicking wrong or whether I should re-pull the dev brunch again
