### 2024-10-29 #7399 @lukepolson
Hi @zapaishchykova , are you able to meet for a brief 10 min google meet call right now so I can see the options you've selected? It could be a weird bug that has to do with not selecting something.
