### 2024-10-30 #7399 @zapaishchykova
Hi, just to keep a paper trail, we met yesterday with @lukepolson discussed a couple of improvements that they will implement into the repo ( loading CT and then switching to other type, more detailed Readme) and I will resume the review once its ready.
Thanks!
