+++ User_Manual.md
@@ -1,6 +1,10 @@
-# PRELUDE: LOADING INPUT IMAGES
+# Module Description
 
-First of all, the input images must be imported into Slicer and loaded in the Slicer scene; the images must be in DICOM format.
+This module enables the reconstruction of raw SPECT projection data, providing customizable options for image modeling and image reconstruction. This document is a user manual for using the extension.
+
+# Loading DICOM Data
+
+The `SlicerSPECTRecon` extension uses DICOM data imported via the DCM data loader of 3D slicer. All data used in reconstruction (e.g. SPECT projection data and CT images for attenuation correction) must first be loaded into 3D slicer as follows:
 
 1. Go to *File* → *Add DICOM Data* or simply click ![dcm](images/dcm.png)
 2. Click *Import DICOM Files*
@@ -8,62 +12,50 @@ First of all, the input images must be imported into Slicer and loaded in the Sl
 4. In the "Dicom Database" on the right hand side, click on the patient name, select the files you want to important, and click "Load" (see Figure 0).
 
  ![importdicom](images/loadfiles.png)
-Figure 0: Loading DICOM images
 
-# DESCRIPTION
+# Extension Instructions
 
-<!-- comments in Markdown language -->
+The purpose of this module is to enable reconstruction of raw SPECT projection data in 3D slicer with a variety of system modeling techniques and algorithms. Similar to reconstruction workflows from clinical software, the extension is composed of multiple steps
 
-This module enables the reconstruction of SPECT projection data, supporting both patient and Simind datasets.
-
-
-<!-- Markdown indentation of nested lists requires 4 spaces  -->
-1. **Data Converters:** Tools for converting projection data and attenuation maps from various formats into DICOM. The module currently supports conversion of the SIMIND data format. Once data has been converted, it will be saved to a DICOM file somewhere on the users computer, which then needs to be loaded via the Slicer DICOM image loader.
-    * *SIMIND Data Converter*: The output of a SIMIND simulation may include multiple energy windows: users specify how many energy windows they want to include in the DICOM file in the "Number of energy windows" section. Since DICOM data is stored in units of counts, and SIMIND data is output in counts/second/MBq (or in some cases counts/second/(MBq/mL)) users must specify a time per projection (second), and scale factor (MBq or MBq/mL) to multiply the SIMIND data by. The noiseless SIMIND data is then sampled according to a Poisson distribution with the given "Random Seed". Users need to select the header files for each window they want to provide (".h00" files); these files must be in the same directory as the data files (".a00" files). Finally, the user must provide a patient name to give to the generated data, and choose a location to save on the computer. To convert a SIMIND attenuation map into DICOM format, users must select the header files (".hct"); this file must be in the same directory as the data file ("ict"). The user then specifies the location on their computer where to save the resulting files.
-
- ![dataconverters](images/converters.png)
-Figure 1.0: Data Converters Panel
-
-2. **Input Data:** Used to obtain photopeak projections from loaded DICOM data. Users can select multiple projection files corresponding to different bed positions of the same scan: the reconstructed object will automatically stitch the separate bed positions together.
+1. **Input Data:** The input data component is used to select the projection data for reconstruction and the corresponding energy window used for reconstruction. Users first select the projection data (previously loaded into slicer) from the drop down menu, and then select the energy window they wish to reconstruct. For uses cases consisting of multiple bed positions (common in some patient scans), multiple projections can be selected; after reconstruction, they are automatically stitched together.
 
  ![inputdata](images/input.png)
-Figure 1.1: Input Data Panel
 
[…]
