### 2024-10-30 #7399 @lukepolson
@zapaishchykova I've updated the user manual [here](https://github.com/PyTomography/slicer_spect_recon/blob/development/User_Manual.md), let me know if there are any issues.
