### 2024-11-01 4d55e67
Filter nodes to display
