### 2024-11-01 e745b07
Fix bug in filtering nodes
