+++ SlicerSPECTRecon/SlicerSPECTRecon.py
@@ -158,37 +158,45 @@ class SlicerSPECTReconWidget(ScriptedLoadableModuleWidget, VTKObservationMixin):
     def filterNMVolumes(self):
         """Filter the projection data to show only Nuclear Medicine volumes."""
         nmSOPClassUID = "1.2.840.10008.5.1.4.1.1.20"  # Standard SOPClassUID for PET images
+        sopclassuidtag = '0008,0016'
         for nodeIndex in range(slicer.mrmlScene.GetNumberOfNodesByClass("vtkMRMLScalarVolumeNode")):
             volNode = slicer.mrmlScene.GetNthNodeByClass(nodeIndex, "vtkMRMLScalarVolumeNode")
             uids = volNode.GetAttribute('DICOM.instanceUIDs')
             if uids is not None:
-                uid = uids.split()[0]
-            filepath = slicer.dicomDatabase.fileForInstance(uid)
-            sopclassuidtag = '0008,0016'
-            sopClassUID = slicer.dicomDatabase.fileValue(filepath, sopclassuidtag)
-            volNode.SetAttribute("SOPClassUID", sopClassUID) # Set attribute in volume with corresponding UID
-            if sopClassUID == nmSOPClassUID:
-                # Rule to filter through the nodes to add
-                self.ui.NM_data_selector.addAttribute("vtkMRMLScalarVolumeNode", "SOPClassUID", nmSOPClassUID)
+                uid = uids.split()
+            if len(uid)>1:
+                filepaths = [slicer.dicomDatabase.fileForInstance(instanceUID) for instanceUID in uid]
+                for filepath in filepaths:
+                    sopClassUID = slicer.dicomDatabase.fileValue(filepath, sopclassuidtag)
+                    volNode.SetAttribute("SOPClassUID", sopClassUID)
+            else:
+                filepath = slicer.dicomDatabase.fileForInstance(uid[0])
+                sopClassUID = slicer.dicomDatabase.fileValue(filepath, sopclassuidtag)
+                volNode.SetAttribute("SOPClassUID", sopClassUID)
+        self.ui.NM_data_selector.addAttribute("vtkMRMLScalarVolumeNode", "SOPClassUID", nmSOPClassUID)
 
     def filterCTVolumes(self):
         """Filter the attenuation data to show only CT volumes."""
         ctSOPClassUID = "1.2.840.10008.5.1.4.1.1.2"  # Standard SOPClassUID for CT images
+        sopclassuidtag = '0008,0016'
         for nodeIndex in range(slicer.mrmlScene.GetNumberOfNodesByClass("vtkMRMLScalarVolumeNode")):
             volNode = slicer.mrmlScene.GetNthNodeByClass(nodeIndex, "vtkMRMLScalarVolumeNode")
-            uids = volNode.GetAttribute('DICOM.instanceUIDs')
-            if uids is not None:
-                uid = uids.split()[0]
-            filepath = slicer.dicomDatabase.fileForInstance(uid)
-            sopclassuidtag = '0008,0016'
-            sopClassUID = slicer.dicomDatabase.fileValue(filepath, sopclassuidtag)
             if volNode.GetAttribute("SOPClassUID"):
                 pass
             else:
-                volNode.SetAttribute("SOPClassUID", sopClassUID) # Set attribute in volume with corresponding UID
-            if sopClassUID == ctSOPClassUID:
-                # Rule to filter through the nodes to add
-                self.ui.attenuationdata.addAttribute("vtkMRMLScalarVolumeNode", "SOPClassUID", ctSOPClassUID) 
+                uids = volNode.GetAttribute('DICOM.instanceUIDs')
+                if uids is not None:
+                    uid = uids.split()
+                if len(uid)>1:
[…]
