### 2024-11-08 a4fb32b
update prior bug no anatomy image
