+++ User_Manual.md
@@ -1,4 +1,46 @@
-# PRELUDE: LOADING INPUT IMAGES
+**DISCLAIMER: This plugin is for academic purposes only and it is not recommended for use in clinical environment.**
+
+# Setup
+
+**Prerequisite**: 
+
+- Latest version and release of 3D Slicer (5.6.1)
+
+## Installation
+
+The *SlicerSPECTRecon* module, when released to the extension manager, shall be installed automatically from *View* -> *Extension manager* -> *Install Extensions* or directly from ![extensions](images/install_extensions.png) in Slicer Home. However, it can also be downloaded or cloned  from [https://github.com/PyTomography/slicer_spect_recon.git](https://github.com/PyTomography/slicer_spect_recon.git). 
+
+**In the case of download**, the insallation proceeds as follows:
+
+1. Download the zip file (see image below) and extract it in a personal directory folder. It will contain, among other things, the folder SlicerSPECTRecon with the SlicerSPECTRecon.py file
+2. Open Slicer and click ![customize](images/customize.png), or alternatively *Edit* -> *Application Settings*. 
+3. In *Modules* -> *Additional module paths*, add the path to the folder containing the module files.
+4. Restart Slicer to apply the changes.
+
+<img src="images/download.png" width="300"/>
+
+**In the case of cloning from GitHub**, the installation proceeds as follows:
+ 
+1. In GitHub copy URL from "Clone with HTTPS" (see screenshot below)
+2. Open terminal in a personal directory folder and execute the line: <pre><code>git clone https://github.com/PyTomography/slicer_spect_recon.git </code></pre>
+(i.e. "git clone" plus the copied URL).
+This will download a folder called *slicer_spect_recon*, with  all the necessary files for the module 
+3. Open Slicer and click ![customize](images/customize.png), or alternatively *Edit* -> *Application Settings*. 
+4. In *Modules* -> *Additional module paths*, add the path to the folder containing the module files (see screenshot below).
+5. Restart Slicer.
+
+<img src="images/clone.png" width="300"/>
+
+<img src="images/modules.png" width="300"/>
+
+Once Slicer is restarted, the SlicerSPECTRecon module will be available from *All Modules* → *Tomographic Reconstruction* → *SlicerSPECTRecon* 
+
+<img src="images/modules_list.png" width="300"/>
+
+## Tutorial Data
+The tutorial data can be obtained [at the following link](https://zenodo.org/records/14172228); it will be used in the subsequent examples in the user manual.
+
+## Data Loading
 
 First of all, the input images must be imported into Slicer and loaded in the Slicer scene; the images must be in DICOM format.
 
@@ -7,103 +49,43 @@ First of all, the input images must be imported into Slicer and loaded in the Sl
 3. Select the folder containing the DICOM files you wish to import, and click "Import" in the file explorer.
 4. In the "Dicom Database" on the right hand side, click on the patient name, select the files you want to important, and click "Load" (see Figure 0).
 
- ![importdicom](images/loadfiles.png)
-Figure 0: Loading DICOM images
+<img src="images/data_loading.png" width="300"/>
 
-# DESCRIPTION
-
-<!-- comments in Markdown language -->
+# Description
 
 This module enables the reconstruction of SPECT projection data, supporting both patient and Simind datasets.
 
-
-<!-- Markdown indentation of nested lists requires 4 spaces  -->
 1. **Data Converters:** Tools for converting projection data and attenuation maps from various formats into DICOM. The module currently supports conversion of the SIMIND data format. Once data has been converted, it will be saved to a DICOM file somewhere on the users computer, which then needs to be loaded via the Slicer DICOM image loader.
[…]
