### 2024-11-17 ba3608f
add loading bar and fix misc changes
