+++ SlicerSPECTRecon/SlicerSPECTRecon.py
@@ -128,11 +128,13 @@ class SlicerSPECTReconWidget(ScriptedLoadableModuleWidget, VTKObservationMixin):
         self._parameterNode = None
         self._updatingGUIFromParameterNode = False
         self._projectionList = None
+        print('init activated')
 
     def setup(self):
         """
         Called when the user opens the module the first time and the widget is initialized.
         """
+        print('setup activated')
         ScriptedLoadableModuleWidget.setup(self)
         # Load widget from .ui file (created by Qt Designer).
         # Additional widgets can be instantiated manually and added to self.layout.
@@ -156,14 +158,14 @@ class SlicerSPECTReconWidget(ScriptedLoadableModuleWidget, VTKObservationMixin):
             ['string', 'string', 'string', 'float', 'float']
         )
 
-    def filterNMVolumes(self):
-        """Filter the projection data to show only Nuclear Medicine volumes."""
-        self.ui.NM_data_selector.noneEnabled = False
-        self.ui.NM_data_selector.addEnabled = False
-        nmSOPClassUID = "1.2.840.10008.5.1.4.1.1.20"  # Standard SOPClassUID for PET images
-        self.ui.NM_data_selector.addAttribute("vtkMRMLScalarVolumeNode", "SOPClassUID", nmSOPClassUID)
-        self.ui.NM_data_selector.setMRMLScene(None)  # Clear the combobox first
-        self.ui.NM_data_selector.setMRMLScene(slicer.mrmlScene)
+    # def filterNMVolumes(self):
+    #     """Filter the projection data to show only Nuclear Medicine volumes."""
+    #     self.ui.NM_data_selector.noneEnabled = False
+    #     self.ui.NM_data_selector.addEnabled = False
+    #     nmSOPClassUID = "1.2.840.10008.5.1.4.1.1.20"  # Standard SOPClassUID for PET images
+    #     self.ui.NM_data_selector.addAttribute("vtkMRMLScalarVolumeNode", "SOPClassUID", nmSOPClassUID)
+    #     self.ui.NM_data_selector.setMRMLScene(None)  # Clear the combobox first
+    #     self.ui.NM_data_selector.setMRMLScene(slicer.mrmlScene)
 
     def filterCTVolumes(self):
         """Filter the attenuation data to show only CT volumes."""
@@ -246,15 +248,19 @@ class SlicerSPECTReconWidget(ScriptedLoadableModuleWidget, VTKObservationMixin):
         
     def cleanup(self) -> None:
         """Called when the application closes and the module widget is destroyed."""
+        print('cleanup activated')
         self.removeObservers()
 
     def enter(self) -> None:
         """Called each time the user opens this module."""
+        print('enter activated')
         # Make sure parameter node exists and observed
-        self.initializeParameterNode()
-        self.filterNMVolumes()
+        #self.initializeParameterNode()
+        self.onMRMLSceneNodeAdded(None,None)
+        #self.filterNMVolumes()
 
     def onMRMLSceneNodeAdded(self, caller, event):
+        print('onMRMLSceneNodeAdded')
         for nodeIndex in range(slicer.mrmlScene.GetNumberOfNodesByClass("vtkMRMLScalarVolumeNode")):
             volNode = slicer.mrmlScene.GetNthNodeByClass(nodeIndex, "vtkMRMLScalarVolumeNode")
             uids = volNode.GetAttribute('DICOM.instanceUIDs')
@@ -270,34 +276,39 @@ class SlicerSPECTReconWidget(ScriptedLoadableModuleWidget, VTKObservationMixin):
                     filepath = slicer.dicomDatabase.fileForInstance(uid[0])
                     sopClassUID = slicer.dicomDatabase.fileValue(filepath, sopclassuidtag)
[…]
