### 2024-11-18 #7399 @ObedDzik
@mstimberg we have made the change to the readme file and moved the data to zenodo. We have also ensured that only the volumes of the required types show up in the drop down of the respective buttons to avoid confusion. We are currently running some tests on this specific change before merging. So @cnmy-ro can go ahead with the review and we will update @zapaishchykova once we merge our changes.
