+++ SlicerSPECTRecon/SlicerSPECTRecon.py
@@ -45,19 +45,13 @@ from Testing.Python.simind_to_dicom import SimindToDicomConverterTest
 
 __submoduleNames__ = [
     "SlicerSPECTReconLogic",
-    "SlicerSPECTReconTest",
-    "vtkkmrmlutils",
-    "simindToDicomConverterTest",
-    "testutils_builder",
-    "transforms",
-    "volumeutils",
-    "systemMatrix",
-    "simindToDicom",
-    "priors",
-    "algorithms",
-    "getmetadatautils",
-    "likelihood",
-    "dicomvalues"
+    "Algorithms",
+    "Callbacks",
+    "MetadataUtils",
+    "Priors",
+    "SimindToDicom",
+    "VolumeUtils",
+    "VtkkmrmlUtils",
 ]
 __package__ = "SlicerSPECTRecon"
 mod = importlib.import_module("Logic", __name__)
@@ -103,10 +97,10 @@ class SlicerSPECTRecon(ScriptedLoadableModule):
         """
         logging.info("\n******* Starting Tests of SlicerSPECTRecon **********\n")
         # Test SIMIND Converter
-        testCase = SimindToDicomConverterTest()
-        testCase.messageDelay = msec
-        testCase.runTest(**kwargs)
-        # # Test Reconstructions
+        # testCase = SimindToDicomConverterTest()
+        # testCase.messageDelay = msec
+        # testCase.runTest(**kwargs)
+        # Test Reconstructions
         testCase = ReconstructionTest()
         testCase.messageDelay = msec
         testCase.runTest(**kwargs)
@@ -128,13 +122,11 @@ class SlicerSPECTReconWidget(ScriptedLoadableModuleWidget, VTKObservationMixin):
         self._parameterNode = None
         self._updatingGUIFromParameterNode = False
         self._projectionList = None
-        print('init activated')
 
     def setup(self):
         """
         Called when the user opens the module the first time and the widget is initialized.
         """
-        print('setup activated')
         ScriptedLoadableModuleWidget.setup(self)
         # Load widget from .ui file (created by Qt Designer).
         # Additional widgets can be instantiated manually and added to self.layout.
@@ -158,14 +150,14 @@ class SlicerSPECTReconWidget(ScriptedLoadableModuleWidget, VTKObservationMixin):
             ['string', 'string', 'string', 'float', 'float']
         )
 
-    # def filterNMVolumes(self):
-    #     """Filter the projection data to show only Nuclear Medicine volumes."""
-    #     self.ui.NM_data_selector.noneEnabled = False
-    #     self.ui.NM_data_selector.addEnabled = False
-    #     nmSOPClassUID = "1.2.840.10008.5.1.4.1.1.20"  # Standard SOPClassUID for PET images
-    #     self.ui.NM_data_selector.addAttribute("vtkMRMLScalarVolumeNode", "SOPClassUID", nmSOPClassUID)
-    #     self.ui.NM_data_selector.setMRMLScene(None)  # Clear the combobox first
-    #     self.ui.NM_data_selector.setMRMLScene(slicer.mrmlScene)
+    def filterNMVolumes(self):
+        """Filter the projection data to show only Nuclear Medicine volumes."""
+        self.ui.NM_data_selector.noneEnabled = False
+        self.ui.NM_data_selector.addEnabled = False
+        nmSOPClassUID = "1.2.840.10008.5.1.4.1.1.20"  # Standard SOPClassUID for PET images
+        self.ui.NM_data_selector.addAttribute("vtkMRMLScalarVolumeNode", "SOPClassUID", nmSOPClassUID)
+        self.ui.NM_data_selector.setMRMLScene(None)  # Clear the combobox first
[…]
