+++ SlicerSPECTRecon/Testing/Python/reconstruction.py
@@ -32,7 +32,7 @@ class ReconstructionTest(ScriptedLoadableModuleTest):
         self.cleanUP()
         logger.info("SlicerSPECTReconTests")
         module_dir = os.path.dirname(__file__)
-        self.resources_dir = os.path.join(module_dir, './../Resources')
+        self.resources_dir = os.path.join(module_dir, './../../Resources')
         self.logic = SlicerSPECTReconLogic()
         try:
             self.delayDisplay("Initializing Dicom Database")
@@ -51,56 +51,45 @@ class ReconstructionTest(ScriptedLoadableModuleTest):
     def runTest(self):
         """Run as few or as many tests as needed here."""
         self.setUp()
-        self.test_load_projection_data()
-        self.test_projection_metadata()
-        self.test_attenuation_map_alignment()
-        self.test_psf_metadata()
-        self.test_scatter()
-        self.run_test_osem()
-        self.run_test_bsrem()
+        self.load_ref_data()
+        # self.test_load_projection_data()
+        # self.test_projection_metadata()
+        # self.test_attenuation_map_alignment()
+        # self.test_psf_metadata()
+        # self.test_scatter()
+        # self.run_test_osem()
+        # self.run_test_bsrem()
         self.run_test_osmaposl()
         self.cleanUP()
-
-    def importDICOMFromURL(self, file_id, folder_name):
-        dicomDir = Path(slicer.app.temporaryPath) /folder_name
-        dicomDir.mkdir(parents=True, exist_ok=True)
-        zip_path = dicomDir / 'dicom.zip'
-        logger.info(f"Downloading {folder_name} DICOM files from drive to {dicomDir}")
-        status = download_file_from_google_drive(file_id, zip_path)
-        if status:
-            with zipfile.ZipFile(zip_path, 'r') as zip_ref:
-                zip_ref.extractall(dicomDir)
-        else:
-            logger.error(f"Failed to download {folder_name} DICOM files")
-            return
-        indexer = ctk.ctkDICOMIndexer()
-        indexer.addDirectory(slicer.dicomDatabase, str(dicomDir))
-        indexer.waitForImportFinished()
-        logger.info(f"Imported DICOM files from drive to database")
-
-    def downloadAndImportNMFiles(self):
-        file_id = "1bCz_hLgASAiQ38QrRlgrJ3lH_lOlqQb1"
-        folder_name = "NM"
-        self.importDICOMFromURL(file_id, folder_name)
-
-    def downloadAndImportCTFiles(self):
-        file_id = "1kbl4nqflovxOB0N9z8YrpmZHzTnPtnyu"
-        folder_name = "CT"
-        self.importDICOMFromURL(file_id, folder_name)
     
+    def load_ref_data(self):
+        url = "https://zenodo.org/records/14190900/files/refs.zip?download=1"
+        data_type = 'ref_data'
+        self.testDir = get_data_from_url(url, data_type)
+        
     def uploadtoDB(self):
         dicomValues = DicomValues()
-        self.downloadAndImportNMFiles()
-        self.downloadAndImportCTFiles()
-        nm_series = slicer.dicomDatabase.seriesForStudy(dicomValues.NM_studyInstanceUID)
-        nm_series_count = len([uid for uid in nm_series if uid in [dicomValues.NM1_seriesInstanceUID, dicomValues.NM2_seriesInstanceUID]])
-        if nm_series_count != 2:
-            raise Exception(f"Expected 2 NM series, but found {nm_series_count}")
-        ct_series = slicer.dicomDatabase.seriesForStudy(dicomValues.CT_studyInstanceUID)
-        ct_series_count = len([uid for uid in ct_series if uid in [dicomValues.CT_seriesInstanceUID]])
[…]
