### 2024-11-20 3c31029
Activate all tests
