### 2024-11-21 6055030
filter nodes and GUI by NM and CT
