+++ SlicerSPECTRecon/SlicerSPECTRecon.py
@@ -4,7 +4,7 @@ import importlib
 def ensure_packages_installed():
     """Ensure required packages are installed."""
     required_packages = {
-        "pytomography": "3.2.4",
+        "pytomography": "3.3.0",
         "beautifulsoup4": None,
     }
     for package, version in required_packages.items():
@@ -127,6 +127,7 @@ class SlicerSPECTReconWidget(ScriptedLoadableModuleWidget, VTKObservationMixin):
         """
         Called when the user opens the module the first time and the widget is initialized.
         """
+        logging.info('setup called')
         ScriptedLoadableModuleWidget.setup(self)
         # Load widget from .ui file (created by Qt Designer).
         # Additional widgets can be instantiated manually and added to self.layout.
@@ -137,11 +138,15 @@ class SlicerSPECTReconWidget(ScriptedLoadableModuleWidget, VTKObservationMixin):
         self.logic = SlicerSPECTReconLogic()
         # Connections
         # # These connections ensure that we update parameter node when scene is closed
+        self.addObserver(slicer.mrmlScene, slicer.vtkMRMLScene.NodeAddedEvent, self.onMRMLSceneNodeAdded)
         self.addObserver(slicer.mrmlScene, slicer.mrmlScene.StartCloseEvent, self.onSceneStartClose)
         self.addObserver(slicer.mrmlScene, slicer.mrmlScene.EndCloseEvent, self.onSceneEndClose)
-        self.addObserver(slicer.mrmlScene, slicer.vtkMRMLScene.NodeAddedEvent, self.onMRMLSceneNodeAdded)
         # These connections ensure that whenever user changes some settings on the GUI, that is saved in the MRML scene
         # (in the selected parameter node).
+        # Add filters to dropdown menus so we only get spect projections / CTs in the dropdown boxes
+        self.ui.NM_data_selector.addAttribute("vtkMRMLScalarVolumeNode", "SOPClassUID", "1.2.840.10008.5.1.4.1.1.20")
+        self.ui.attenuationdata.addAttribute("vtkMRMLScalarVolumeNode", "SOPClassUID", "1.2.840.10008.5.1.4.1.1.2")
+        self.ui.anatomyPriorImageNode.addAttribute("vtkMRMLScalarVolumeNode", "SOPClassUID", "1.2.840.10008.5.1.4.1.1.2")
         self.setupConnections()
         # initialize for loading data in case
         # Add uncertainty table
@@ -149,40 +154,9 @@ class SlicerSPECTReconWidget(ScriptedLoadableModuleWidget, VTKObservationMixin):
             ['Image', 'Segmentation', 'Mask', 'Absolute Uncertainty', 'Percent Uncertainty'],
             ['string', 'string', 'string', 'float', 'float']
         )
-
-    def filterNMVolumes(self):
-        """Filter the projection data to show only Nuclear Medicine volumes."""
-        self.ui.NM_data_selector.noneEnabled = False
-        self.ui.NM_data_selector.addEnabled = False
-        nmSOPClassUID = "1.2.840.10008.5.1.4.1.1.20"  # Standard SOPClassUID for PET images
-        self.ui.NM_data_selector.addAttribute("vtkMRMLScalarVolumeNode", "SOPClassUID", nmSOPClassUID)
-        self.ui.NM_data_selector.setMRMLScene(None)  # Clear the combobox first
-        self.ui.NM_data_selector.setMRMLScene(slicer.mrmlScene)
-
-    def filterCTVolumes(self):
-        """Filter the attenuation data to show only CT volumes."""
[…]
