+++ paper/paper.bib
@@ -1,0 +1,174 @@
+@ARTICLE{osem,
+  author={Hudson, H.M. and Larkin, R.S.},
+  journal={IEEE Transactions on Medical Imaging}, 
+  title={Accelerated image reconstruction using ordered subsets of projection data}, 
+  year={1994},
+  volume={13},
+  number={4},
+  pages={601-609},
+  doi={10.1109/42.363108}}
+
+@ARTICLE{OSL,
+  author={Green, P.J.},
+  journal={IEEE Transactions on Medical Imaging}, 
+  title={Bayesian reconstructions from emission tomography data using a modified EM algorithm}, 
+  year={1990},
+  volume={9},
+  number={1},
+  pages={84-93},
+  doi={10.1109/42.52985}}
+
+@article{simind,
+title = {A Monte Carlo program for the simulation of scintillation camera characteristics},
+journal = {Computer Methods and Programs in Biomedicine},
+volume = {29},
+number = {4},
+pages = {257-272},
+year = {1989},
+issn = {0169-2607},
+doi = {10.1016/0169-2607(89)90111-9},
+url = {https://www.sciencedirect.com/science/article/pii/0169260789901119},
+author = {Michael Ljungberg and Sven-Erik Strand},
+keywords = {Monte Carlo, SPECT, Scatter simulation, Photon transport, Pile-up, Imaging, Spectrum, Scintillation camera},
+abstract = {There is a need for mathematical modelling for the evaluation of important parameters for photon imaging systems. A Monte Carlo program which simulates medical imaging nuclear detectors has been developed. Different materials can be chosen for the detector, a cover and a phantom. Cylindrical, spherical, rectangular and more complex phantom and source shapes can be simulated. Photoelectric, incoherent, coherent interactions and pair production are simulated. Different detector parameters, e.g. the energy pulse-height distribution and pulse pile-up due to finite decay time of the scintillation light emission, can be calculated. An energy resolution of the system is simulated by convolving the energy imparted with an energy-dependent Gaussian function. An image matrix of the centroid of the events in the detector can be simulated. Simulation of different collimators permits studies on spatial resolution and sensitivyt. Comparisons of our results with experimental data and other published results have shown good agreement. The usefulness of the Monte Carlo code for the accurately simulation of important parameters in scintillation camera systems, stationary as well as SPECT (single-photon emission computed tomography) systems, has been demonstrated.}
+}
+
+@ARTICLE{RDP,
+  author={Nuyts, J. and Beque, D. and Dupont, P. and Mortelmans, L.},
+  journal={IEEE Transactions on Nuclear Science}, 
+  title={A concave prior penalizing relative differences for maximum-a-posteriori reconstruction in emission tomography}, 
+  year={2002},
+  volume={49},
+  number={1},
+  pages={56-60},
+  doi={10.1109/TNS.2002.998681}}
+
+@INPROCEEDINGS{STIR,
+  author={Thielemans, Kris and Mustafovic, Sanida and Tsoumpas, Charalampos},
+  booktitle={2006 IEEE Nuclear Science Symposium Conference Record}, 
+  title={STIR: Software for Tomographic Image Reconstruction Release 2}, 
+  year={2006},
+  volume={4},
+  number={},
+  pages={2174-2176},
+  doi={10.1109/NSSMIC.2006.354345}}
+
+@article{Ashrafinia_2017,
+doi = {10.1088/1361-6560/aa6911},
+url = {https://dx.doi.org/10.1088/1361-6560/aa6911},
+year = {2017},
+month = {may},
+publisher = {IOP Publishing},
+volume = {62},
+number = {12},
+pages = {5149},
[…]
