+++ paper/paper.md
@@ -1,0 +1,81 @@
+---
+title: 'SlicerSPECTRecon: A 3D Slicer Extension for SPECT Image Reconstruction'
+tags:
+  - 3D slicer
+  - nuclear medicine
+  - tomography
+  - spect
+  - image reconstruction
+authors:
+  - name: Obed K. Dzikunu
+    orcid: 0000-0002-1122-0629
+    corresponding: true # (This is how to denote the corresponding author)
+    affiliation: "2, 3" # (Multiple affiliations must be quoted)
+  - name: Maziar Sabouri
+    affiliation: "1, 2"
+  - name: Shadab Ahamed
+    affiliation: "1, 2"
+  - name: Carlos Uribe
+    affiliation: "2, 4, 5"
+  - name: Arman Rahmim
+    affiliation: "1, 2, 3, 5"
+  - name: Luke Polson
+    affiliation: "1, 2"    
+affiliations:
+ - name: Deparment of Physics & Astronomy, University of British Columbia, Vancouver Canada
+   index: 1
+ - name: Department of Integrative Oncology, BC Cancer Research Institute, Vancouver Canada
+   index: 2
+ - name: School of Biomedical Engineering, University of British Columbia, Vancouver Canada
+   index: 3
+ - name: Department of Radiology, University of British Columbia, Vancouver Canada
+   index: 4
+ - name: Molecular Imaging and Therapy Department, BC Cancer, Vancouver Canada
+   index: 5
+date: 06 September 2024
+bibliography: paper.bib
+
+# Optional fields if submitting to a AAS journal too, see this blog post:
+# https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
+aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
+aas-journal: Astrophysical Journal <- The name of the AAS journal.
+---
+
+# Summary
+
+`SlicerSPECTRecon` is a `3D Slicer` [@Kikinis2014-3DSlicer] extension designed for Single Photon Emission Computed Tomography (SPECT) image reconstruction. It offers a range of popular reconstruction algorithms and requires raw projection data from clinical or Monte Carlo simulated scanners. Built with the PyTomography python library [@pytomography], it features GPU-accelerated functionality for fast reconstruction. The extension includes a graphical user interface for the selection of reconstruction parameters, and reconstructed images can be post-processed using all available `3D Slicer` functionalities.
+
+
+# Statement of need
+
+SPECT imaging is used to measure the 3D distribution of a radioactive molecule within a patient; it requires (i) acquisition of 2D `projection` images at different angles using a gamma camera followed by (ii) the use of a tomographic image reconstruction algorithm to obtain a 3D radioactivity distribution consistent with the acquired data. While there exist many commercial products for image reconstruction, as well as multiple low-level software packages [@pytomography] [@STIR] [@castor], there is currently no open source graphic user interface (GUI) that provides access to the latest reconstruction algorithms. Due to continuing research interest in the implications of different reconstruction techniques on various clinical tasks [@hermes_spect_ap] [@spect_ai], there is a need for a open-source GUI for SPECT reconstruction.
+
+
+# Overview of SlicerSPECTRecon
+
+Typical reconstruction algorithms in nuclear medicine attempt to maximize a likelihood function via
+\begin{equation}
+  \hat{x} = \underset{x;A}{\text{argmax}}~L(y|x,H,s)
+\end{equation}
[…]
