### 2024-11-21 #7399 @lukepolson
@cnmy-ro I should also note that although the references don't appear explicitly in the markdown file as you commented, running the compile code to generate the PDF of the paper does add them to the file. @zapaishchykova don't hesitate to contact if you have any issues, and be sure to checkout the youtube tutorial linked in the User_Manual : )
