### 2024-11-22 #7399 @zapaishchykova
![Screenshot 2024-11-22 at 10 27 19 AM](https://github.com/user-attachments/assets/0cff02f0-5362-4fe6-9376-c52cd26edf88)

Hello @lukepolson, qq: I don't see n.6 from https://github.com/PyTomography/slicer_spect_recon/blob/main/User_Manual.md (see attached screenshot).
