### 2024-11-22 #7399 @lukepolson
Hi @zapaishchykova , good catch! : ) I've updated the manual to consider the case for "Joint Multi-Photopeak Reconstruction." For the example data, I go over that in the youtube tutorial I've linked in the user manual. I don't really specify any of the parameters for the example data in the user manual as it's meant to be more of a general reference, but rather defer to the video for tutorial specifics. The reasoning for this is that we reconstruct the data using many different possible configurations in the video tutorial, so there is no one correct way to do it. However, if you have a strong objection to this, then I can pursue adding this to the user manual.
