### 2024-11-22 #7399 @zapaishchykova
As a suggestion, you can explicitly guide users to your video for guidance.
I have no further comments.
