### 2024-11-25 #7399 @mstimberg
Many thanks again for your reviews @cnmy-ro and @zapaishchykova. With both reviewers recommending acceptance of the paper/software in its current state, I will now proceed with the final steps before handing things over to a topic editor for the actual publication. @lukepolson please have a look at the "author tasks" in the checklist below and comment here when you've taken care of them.
