### 2024-11-25 #7399 @lukepolson
Hi @mstimberg, I have complete the checklist: namely:
* I have double checked the authors and affiliations, and I have added the corresponding ORCID IDs for all authors and modified the paper
* I have made a release of the software on [github](https://github.com/PyTomography/slicer_spect_recon/releases/tag/v1.0.0) under release v1.0.0
* I have archived the release on [zenodo](https://zenodo.org/records/14218029) and the DOI is 10.5281/zenodo.14218029)
* On Zenodo, I have made sure that the title and authors list (including ORCIDs) match those in the JOSS paper and are in the same order. These are listed as "contributors" on Zenodo (I only added myself as "creator" since this corresponds to creation of the archive),
* I have made sure that the license listed on Zenodo matches that of the paper, e.g. the MIT license.

Thank you for your help here : ). By the way, we are still waiting for the official PyTomography paper to be published, which should be completed in the next week or two. Is it possible, after publication with JOSS, to change the PyTomography paper reference from arxiv to SoftwareX?
