### 2024-11-28 32f8951
paper: remove AAS fields
