+++ paper/paper.md
@@ -12,15 +12,20 @@ authors:
     corresponding: true # (This is how to denote the corresponding author)
     affiliation: "2, 3" # (Multiple affiliations must be quoted)
   - name: Maziar Sabouri
+    orcid: 0000-0001-7525-8952
     affiliation: "1, 2"
   - name: Shadab Ahamed
+    orcid: 0000-0002-2051-6085
     affiliation: "1, 2"
   - name: Carlos Uribe
+    orcid: 0000-0003-3127-7478
     affiliation: "2, 4, 5"
   - name: Arman Rahmim
+    orcid: 0000-0002-9980-2403
     affiliation: "1, 2, 3, 5"
   - name: Luke Polson
-    affiliation: "1, 2"    
+    orcid: 0000-0002-3182-2782
+    affiliation: "1, 2" 
 affiliations:
  - name: Deparment of Physics & Astronomy, University of British Columbia, Vancouver Canada
    index: 1
