### 2024-11-28 97a90da
paper: normalize format for affiliations
