### 2024-11-28 2cfbb22
paper: minor capitalization fixes
