+++ paper/paper.bib
@@ -19,7 +19,7 @@
   doi={10.1109/42.52985}}
 
 @article{simind,
-title = {A Monte Carlo program for the simulation of scintillation camera characteristics},
+title = {A {Monte Carlo} program for the simulation of scintillation camera characteristics},
 journal = {Computer Methods and Programs in Biomedicine},
 volume = {29},
 number = {4},
@@ -125,7 +125,7 @@ doi={10.1007/978-1-4614-7657-3_19},
 }
 
 @misc{pytomography,
-      title={PyTomography: A Python Library for Quantitative Medical Image Reconstruction}, 
+      title={PyTomography: A {Python} Library for Quantitative Medical Image Reconstruction}, 
       author={Lucas Polson and Roberto Fedrigo and Chenguang Li and Maziar Sabouri and Obed Dzikunu and Shadab Ahamed and Nikolaos Karakatsanis and Arman Rahmim and Carlos Uribe},
       year={2024},
       eprint={2309.01977},
