### 2024-11-28 #7399 @mstimberg
@lukepolson I proposed a few minor edits in a PR here: PyTomography/slicer_spect_recon#29, please merge if you agree with them.
