### 2024-11-29 #7399 @lukepolson
@mstimberg I have merged all your edits into both the `main` and `j_paper` branches. (My apologies about the ORCID IDs, I had added them in my own fork but forgot to merge to both the `main` and the `jpaper` branches of the primary repository)
