### 2024-12-03 #7399 @Kevin-Mattheus-Moerman
- [ ] Please ensure that the author set on the archive matches the paper author set. You do not need to create a new archive, you can simply edit the current one manually.
- [ ] Check `Deparment` in  your affiliations, that should likely read `Department`
- [ ] `seperate` should read `separate`
- [ ] `overlayed` should likely read `overlaid`
