### 2024-12-03 #7399 @ObedDzik
@Kevin-Mattheus-Moerman thanks for the catch. I just made the corrections. I think for the author list the issue was with my middle name, so I took that out of the paper itself. Now, it's just the first and last name (as in the archive). All the other typos have been corrected.
