### 2024-12-04 #7399 @Kevin-Mattheus-Moerman
@ObedDzik I cannot see the changes to the paper. Have you made these in a particular branch?
