### 2024-12-04 #7399 @ObedDzik
@Kevin-Mattheus-Moerman We have made the change the version tag on zenodo. But regarding the corrections, I made the changes on the main branch, and also merged with the j_paper branch. From the draft paper that the editorial bot generated, I can see the deparment is now department and overlayed is overlaid...unless there's something else I am not catching.
