### 2024-12-04 #7399 @lukepolson
@Kevin-Mattheus-Moerman thanks for the help! Regarding the names, it seems that Zenodo does not allow middle initials, though it would be preferable to have this on the JOSS paper. When matchning the names on the Zenodo archive, is it required that we remove these in the JOSS paper?
