### 2024-12-04 #7399 @lukepolson
@Kevin-Mattheus-Moerman we have modified the main branch with the spelling corrections you have found as well here https://github.com/PyTomography/slicer_spect_recon/blob/main/paper/paper.md (in the main branch) and also merged to the j_paper branch
