### 2024-12-10 #7399 @lukepolson
@mstimberg I hope its not too late, but we just realized that the .png used in the paper was an old one; completely our fault overlooking this. Is it possible to regenerate and re-upload?
