### 2024-12-13 #7399 @Kevin-Mattheus-Moerman
@lukepolson @mstimberg yes we can fix this. Have you updated the figure already? If so then I can update the paper.
