### 2025-01-04 310bf48
add UI for advanced PSF modeling
