+++ paper/compile.sh
@@ -1,5 +1,1 @@
-docker run --rm \
-    --volume $PWD:/data \
-    --user $(id -u):$(id -g) \
-    --env JOURNAL=joss \
-    openjournals/inara
+docker run --rm --volume $PWD:/data --user $(id -u):$(id -g) --env JOURNAL=joss openjournals/inara
