### 2025-01-04 #7399 @lukepolson
Hi @Kevin-Mattheus-Moerman, sorry for the late reply. The picture has indeed been updated. I also added a new reference to the official PyTomography article (the underlying code of this library), which was also recently published.
