### 2025-01-09 #7399 @Kevin-Mattheus-Moerman
@lukepolson let me know if you are now happy with that image so we can close this. Thanks
