### 2025-01-09 #7399 @lukepolson
@Kevin-Mattheus-Moerman My apologies again, I just realized that the change of reference to the PyTomography paper was in the `main` and not the `j_paper` branch. I just merged `main` into `j_paper` so all the references should now be up to date. Sorry for the inconveniences!!
